backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "pakupek",email = "adam.akimowicz13@onet.eu"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "asgiref (==3.8.1)",
    "django (==4.2.20)",
    "django-axes (==7.0.2)",
    "sqlparse (==0.5.3)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.2)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
